refactor(query): Break apart query parsing for unit testing
