+++ src/pages/[...parts].tsx
@@ -3,7 +3,7 @@ import DiffFiles from "components/Diff/DiffFiles";
 import Layout from "components/Layout";
 import Loading from "components/Loading";
 import destination from "lib/destination";
-import parseQuery from "lib/utils/parse-query";
+import parseQuery from "lib/query";
 import rawQuery from "lib/utils/raw-query";
 import setCacheControl from "lib/utils/set-cache-control";
 import specsToDiff from "lib/utils/specs-to-diff";
